chore: refactoring data-plane packages
