+++ extensions/data-plane-ionos-s3/src/test/java/com/ionos/edc/dataplane/ionos/s3/datasource/IonosDataSourceTest.java
@@ -1,12 +1,11 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package com.ionos.edc.dataplane.ionos.s3;
+package com.ionos.edc.dataplane.ionos.s3.datasource;
 
 import com.ionos.edc.extension.s3.connector.S3Connector;
 import com.ionos.edc.extension.s3.types.S3Object;
 import org.eclipse.edc.connector.dataplane.spi.pipeline.StreamFailure;
-import org.eclipse.edc.spi.monitor.Monitor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
@@ -34,12 +33,9 @@ class IonosDataSourceTest {
 
     private static final String TEST_FOLDER_NAME = "devices/";
     private static final String TEST_SUB_FOLDER_1_NAME = "device1/";
-    private static final String TEST_SUB_FOLDER_2_NAME = "device2/";
 
     @Mock
     private S3Connector s3Connector;
-    @Mock
-    private Monitor monitor;
 
     @BeforeEach
     public void setup() {
@@ -54,7 +50,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FILE_1_NAME)
                 .build();
@@ -74,7 +69,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FILE_1_NAME)
                 .build();
@@ -103,7 +97,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .build();
@@ -144,7 +137,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .filterIncludes("device[1-2]-data.csv")
@@ -182,7 +174,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .filterExcludes("device[1-2]-data.csv")
@@ -217,7 +208,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .build();
@@ -263,7 +253,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .filterIncludes("device[1-2]-data.csv")
@@ -302,7 +291,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .filterExcludes("device1-data.csv")
@@ -346,7 +334,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .filterIncludes("device1/device[3-4]-data.csv")
@@ -385,7 +372,6 @@ class IonosDataSourceTest {
 
         var dataSource = IonosDataSource.Builder.newInstance()
                 .client(s3Connector)
-                .monitor(monitor)
                 .bucketName(TEST_BUCKET)
                 .blobName(TEST_FOLDER_NAME)
                 .filterExcludes("device1/device4-data.csv")
